requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic[email]
openpyxl